Force Jenkins to see xunit-file module
For some reason (I think a slightly different npm version), transient
dependencies are behaving slightly differently on jenkinz. Though mocha and
xunit are installed, the mocha that gulp-mocha requires can't see xunit-file.
If the gulpfile could import xunit-file directly that would bring it into scope
(modules are cached), but mocha is not a direct requirement.

So, add mocha as a direct dev requirement, and include xunit-file in the
gulpfile before we invoke gulp-mocha