VBID-3250 Editable component
for making a DOM element click-to-edit style.

takes a value and onChange props, assumes that there's an underlying
<input> you want to wrap.  listens for onKeyDown and onBlur events from
this wrapped element, calling the onChange prop with the value of the element

the "value" prop will be whatever value you want to display, allowing
for formatting, etc

stole the "Editable" name for this component.  renamed previous
"Editable" to "Editable.Element"

in Editable.Element:
  - added props.active so we could keep the little editing styling jawn
    while the element was being editing, in the event the user
    mouseouts.
  - `#isOpened` takes props.active into account as well

Change-Id: I5bbacdbb648f65496ac0a0a4b4d2c93ddd066960